ctf_project/requirements.txt
django==3.2
Pillow==8.2.0
django-mysql==3.11.1
django-crispy-forms==1.11.2
django-cleanup==5.2.0